Reinitialize cyclic barrier only once per cycle
It was done for all threads syncing on a barrier. The one reaching the limit but
also all waiting threads. So it could be reset by a thread while others have
already started another round, leading to an deadlock (because some increments
will be lost then).

Now, cyclic barrier is reset for a new round by the thread reaching the limit
and only this one.