Don't expect payload on 304-Not-Modified response
304-Not-Modified responses must not contain content or trailers and are
terminated at the end of the header section. So don't try to read any
message payload.

This patch fixes a bug when a 304 response contains a Content-Length or a
Transfer-Encoding header. In this case, and without this patch, VTest tries
to honor these headers when found and waits for a body.