varnishtest: Add include command
The contents of a file included by this command are executed as-is. The
command is global, which means that the include command can be used
inside HTTP specs as the test case shows. We require at least one
argument, and subsequent file path arguments are parsed in order.

This command can be used to build more complex VTCs, as well as reducing
repetitiveness. We make no attempts to avoid circular includes.