Makefile: build C files individually (and possibly in parallel)
The current Makefile has a limitation that affects small systems which
is that all the C files are built at once as a single command, there's
no .o file produced. First, this means that there's no possibility of
parallel builds, which is quite noticeable on small many-core low-freq
machines, but the most annoying is that when an error happens, it can
be after quite some time and it takes equally long to verify if a fix
works (e.g. just a missing lib or pcre2.h that fails at the very end
in vre.c).

This patch attempts to solve this while staying compatible with both
BSD make and GNU make. For this it just enumerates all .o files
corresponding to their .c counterpart from a simple shell command, and
issues a recursive compile-only make for them before the final linking
stage. We indeed don't have any portable globbing possibility at the
makefile level that covers both flavors.

Regardless the approach works remarkably well and allows to restart a
failed build, as well as to parallelize objects using -j. Now a missing
pcre2.h error that would previously trigger after 4mn18 is detected at
2m29 the first time, and in only 3.7s for the new attempt, thus saving
4.15mn for the second attempt, which is a significant improvement for
the user trying to solve dependencies.

The FLAGS variable would contain both CFLAGS and LDFLAGS (-s) so it was
kept without the default -s and a pair of CFLAGS & LDFLAGS was created
from it so that it continues to work with unmodified scripts. This allows
to drop -s during the compile-only phase as some compilers do not like
it. Finally, the "clean" rule was updated to also remove object files
(which previously did not exist).